orion
Module hazardous

[Caution] Low-level API.

Caution:

Usage of the hazardous module is only intended for advanced users. hazardous contains implementations with a much higher degree of control. It is also much easier to misuse those implementations. Only use hazardous if absolutely necessary.

Modules

aead

AEADs (Authenticated Encryption with Associated Data).

ecc

Elliptic-Curve Cryptography.

hash

Cryptographic hash functions.

hpke

Hybrid Public Key Encryption (HPKE).

kdf

KDFs (Key Derivation Function) and PBKDFs (Password-Based Key Derivation Function).

kem

Key Encapsulation Mechanisms (KEMs).

mac

MACs (Message Authentication Code).

stream

Stream ciphers.