orion::hazardous::aead
Module chacha20poly1305

AEAD ChaCha20Poly1305 as specified in the RFC 8439.

Parameters:

secret_key: The secret key.

nonce: The nonce value.

ad: Additional data to authenticate (this is not encrypted and can be None).

ciphertext_with_tag: The encrypted data with the corresponding 16 byte Poly1305 tag appended to it.

plaintext: The data to be encrypted.

dst_out: Destination array that will hold the ciphertext_with_tag/plaintext after encryption/decryption.

ad: “A typical use for these data is to authenticate version numbers, timestamps or monotonically increasing counters in order to discard previous messages and prevent replay attacks.” See libsodium docs for more information.

nonce: “Counters and LFSRs are both acceptable ways of generating unique nonces, as is encrypting a counter using a block cipher with a 64-bit block size such as DES. Note that it is not acceptable to use a truncation of a counter encrypted with block ciphers with 128-bit or 256-bit blocks, because such a truncation may repeat after a short time.” See RFC for more information.

Errors:

An error will be returned if:

The length of dst_out is less than plaintext + POLY1305_OUTSIZE when calling seal().

The length of dst_out is less than ciphertext_with_tag - POLY1305_OUTSIZE when calling open().

The length of ciphertext_with_tag is not at least POLY1305_OUTSIZE.

The received tag does not match the calculated tag when calling open().

plaintext.len() + POLY1305_OUTSIZE overflows when calling seal().

Converting usize to u64 would be a lossy conversion.

Panics:

A panic will occur if:

More than 2^32-1 * 64 bytes of data are processed.

Security:

It is critical for security that a given nonce is not re-used with a given key. Should this happen, the security of all data that has been encrypted with that given key is compromised.

Only a nonce for XChaCha20Poly1305 is big enough to be randomly generated using a CSPRNG.

To securely generate a strong key, use SecretKey::generate().

The length of the plaintext is not hidden, only its contents.

Recommendation:

It is recommended to use XChaCha20Poly1305 when possible.

Example:

use orion::hazardous::aead;

let secret_key = aead::chacha20poly1305::SecretKey::generate();

// WARNING: This nonce is only meant for demonstration and should not
// be repeated. Please read the security section.
let nonce = aead::chacha20poly1305::Nonce::from([0u8; 12]);
let ad = "Additional data".as_bytes();
let message = "Data to protect".as_bytes();

// Length of the above message is 15 and then we accommodate 16 for the Poly1305
// tag.

let mut dst_out_ct = [0u8; 15 + 16];
let mut dst_out_pt = [0u8; 15];
// Encrypt and place ciphertext + tag in dst_out_ct
aead::chacha20poly1305::seal(&secret_key, &nonce, message, Some(&ad), &mut dst_out_ct)?;
// Verify tag, if correct then decrypt and place message in dst_out_pt
aead::chacha20poly1305::open(&secret_key, &nonce, &dst_out_ct, Some(&ad), &mut dst_out_pt)?;

assert_eq!(dst_out_pt.as_ref(), message.as_ref());

Re-exports

pub use crate::hazardous::stream::chacha20::Nonce;

pub use crate::hazardous::stream::chacha20::SecretKey;

Functions

open

AEAD ChaCha20Poly1305 decryption and authentication as specified in the RFC 8439.

seal

AEAD ChaCha20Poly1305 encryption and authentication as specified in the RFC 8439.