orion::hazardous
Module hpke

Hybrid Public Key Encryption (HPKE).

Structs

DHKEM_X25519_SHA256_CHACHA20

HPKE suite: DHKEM(X25519, HKDF-SHA256), HKDF-SHA256 and ChaCha20Poly1305.

ModeAuth

HPKE Auth mode. Encrypt data to a public key with sender authentication.

ModeAuthPsk

HPKE AuthPsk mode. Encrypt data to a public key, with sender authentication and an additional preshared-key.

ModeBase

HPKE Base mode. Encrypt data to a public key, without sender authentication.

ModePsk

HPKE Psk mode. Encrypt data to a public key, using a preshared-key providing sender authentication.

Enums

Role

The role for an instance of HPKE mode.