orion::hazardous::hpke
Struct DHKEM_X25519_SHA256_CHACHA20

pub struct DHKEM_X25519_SHA256_CHACHA20 { /* private fields */ }

HPKE suite: DHKEM(X25519, HKDF-SHA256), HKDF-SHA256 and ChaCha20Poly1305.

Note about serialized private keys for this suite

RFC 9180 defines the format of X25519 serialized private keys as the clamped version. According to the standard, (de)serializing from/to a private key requires clamping input/output. This implementation adheres to this requirement, and as such, calling unprotected_as_bytes() on the private key used with this suite will return its clamped version.

The original RFC 9180 test vectors for this suite do on the contrary not include this clamping, so if someone were to compare or otherwise use the output of unprotected_as_bytes(), and expect it to be equal that of other implementations, it might not be. This does not affect interoperability in any other way, meaning HPKE data encrypted with Orion will still decrypt successfully with different HPKE implementations.

The test-vector issues have been reported: https://www.rfc-editor.org/errata/eid7121, https://github.com/cfrg/draft-irtf-cfrg-hpke/issues/255

Implementations

impl DHKEM_X25519_SHA256_CHACHA20

pub const KEM_CT_SIZE: usize = 32usize

Size of the HPKE suite KEM ciphertext/encapsulated key.

pub const KEM_SS_SIZE: usize = 32usize

Size of the HPKE suite KEM shared secret.

pub const VERSION_ID: &[u8; 7] = b"HPKE-v1"

Version identifier for this HPKE scheme.

pub const HPKE_ID: &[u8; 4] = b"HPKE"

HPKE ID for this HPKE scheme.

pub const KEM_ID: [u8; 2] = _

KEM ID for this HPKE scheme’s KEM (in LE bytes).

pub const KDF_ID: [u8; 2] = _

KDF ID for this HPKE scheme’s KDF (in LE bytes).

pub const AEAD_ID: [u8; 2] = _

AEAD ID for this HPKE scheme’s AEAD (in LE bytes).

pub const EXPORT_SECRET_MAXLEN: usize = 8_160usize

The maximum length of export secret that may be requested.

pub const NN: usize = 12usize

Nonce size for this suite’s AEAD (https://www.rfc-editor.org/rfc/rfc9180.html#section-7.3).

pub const NH: usize = 32usize

Output size for this suite’s KDF (https://www.rfc-editor.org/rfc/rfc9180.html#section-7.2).