orion::hazardous::hpke
Struct ModePsk

pub struct ModePsk<S> { /* private fields */ }

HPKE Psk mode. Encrypt data to a public key, using a preshared-key providing sender authentication.

Parameters:

pubkey_r: Public key of the recipient.

info: Optional additional information for HPKE state. Must be identical for sender and recipient.

enc: Encapsulated key, generated by the sender.

secret_key_r: Secret key of the recipient.

exporter_context: Contextual information used during secret export.

psk: Pre-shared key known to both sender and recipient.

psk_id: The identifier for psk.

secret_ephemeral: Ephemeral private key for deterministic encapsulation.

Errors:

An error will be returned if:

info is longer than 64 bytes

out buffer is longer than S::EXPORT_SECRET_MAXLEN when exporting secrets with Self::export_secret()

exporter_context is longer than 64 bytes

The internal counter reaches u64::MAX and a call to Self::seal()/Self::open() is made

Calling Self::seal() when the role is Role::Recipient

Calling Self::open() when the role is Role::Sender

Calling Self::open() on a set of messages that does not match the order of how they were Self::seal()’ed (re-ordering)

psk or psk_id are empty

psk is less than 32 bytes or more than 64 bytes

psk_id is more than 64 bytes

If a shared secret is all-zero.

If ikm.len() < 32 when calling derive_keypair() on a suite’s KEM.

Panics:

A panic will occur if:

getrandom panics during Self::new_sender() or Self::psk_seal().

Security:

When deriving a keypair deterministically instead of generating it randomly, the input ikm must have at least as much entropy as the security level that is desired. For DHKEM_X25519_SHA256_CHACHA20 this means 256 bits.

The ikm used as input for derive_keypair() must never be reused.

The secret_ephemeral must never be reused.

Example:

use orion::hazardous::hpke::{ModePsk, DHKEM_X25519_SHA256_CHACHA20};
use orion::hazardous::kem::x25519_hkdf_sha256::DhKem;

let (sender_secret, sender_public) = DhKem::generate_keypair()?;
let (recipient_secret, recipient_public) = DhKem::generate_keypair()?;
let psk = b"any preshared secret key that is 32 bytes minimum";
let psk_id = b"identifier for psk";

// Streaming-based API
let mut aead_ct_out0 = [0u8; 32];
let mut aead_ct_out1 = [0u8; 32];
let mut aead_ct_out2 = [0u8; 32];

let (mut hpke_sender, enc) = ModePsk::<DHKEM_X25519_SHA256_CHACHA20>::new_sender(&recipient_public, b"info parameter", psk, psk_id)?;
hpke_sender.seal(&[0u8; 16], b"aad parameter 0", &mut aead_ct_out0)?;
hpke_sender.seal(&[1u8; 16], b"aad parameter 1", &mut aead_ct_out1)?;
hpke_sender.seal(&[2u8; 16], b"aad parameter 2", &mut aead_ct_out2)?;

let mut hpke_recipient = ModePsk::<DHKEM_X25519_SHA256_CHACHA20>::new_recipient(&enc, &recipient_secret, b"info parameter", psk, psk_id)?;
let mut aead_pt_out0 = [0u8; 16];
let mut aead_pt_out1 = [0u8; 16];
let mut aead_pt_out2 = [0u8; 16];
hpke_recipient.open(&aead_ct_out0, b"aad parameter 0", &mut aead_pt_out0)?;
hpke_recipient.open(&aead_ct_out1, b"aad parameter 1", &mut aead_pt_out1)?;
hpke_recipient.open(&aead_ct_out2, b"aad parameter 2", &mut aead_pt_out2)?;

assert_eq!(&aead_pt_out0, &[0u8; 16]);
assert_eq!(&aead_pt_out1, &[1u8; 16]);
assert_eq!(&aead_pt_out2, &[2u8; 16]);

// One-shot API
let enc = ModePsk::<DHKEM_X25519_SHA256_CHACHA20>::psk_seal(&recipient_public, b"info parameter", psk, psk_id, &[3u8; 16], b"aad", &mut aead_ct_out0)?;
ModePsk::<DHKEM_X25519_SHA256_CHACHA20>::psk_open(&enc, &recipient_secret, b"info parameter", psk, psk_id, &aead_ct_out0, b"aad", &mut aead_pt_out0)?;
assert_eq!(&aead_pt_out0, &[3u8; 16]);

Implementations

impl<S> ModePsk<S>

pub const MODE_ID: u8 = 1u8

HPKE Psk mode ID.

impl<S: Suite + Psk> ModePsk<S>

pub fn new_sender( pubkey_r: &S::PublicKey, info: &[u8], psk: &[u8], psk_id: &[u8], ) -> Result<(Self, S::EncapsulatedKey), UnknownCryptoError>

Available on crate feature safe_api only.

HPKE Psk mode sender.

pub fn new_sender_deterministic( pubkey_r: &S::PublicKey, info: &[u8], psk: &[u8], psk_id: &[u8], secret_ephemeral: S::PrivateKey, ) -> Result<(Self, S::EncapsulatedKey), UnknownCryptoError>

HPKE Psk mode sender with a supplied ephemeral private key, which is taken ownership of.

pub fn new_recipient( enc: &S::EncapsulatedKey, secret_key_r: &S::PrivateKey, info: &[u8], psk: &[u8], psk_id: &[u8], ) -> Result<Self, UnknownCryptoError>

HPKE Psk mode recipient.

pub fn seal( &mut self, plaintext: &[u8], aad: &[u8], out: &mut [u8], ) -> Result<(), UnknownCryptoError>

Context-aware sealing operations.

pub fn open( &mut self, ciphertext: &[u8], aad: &[u8], out: &mut [u8], ) -> Result<(), UnknownCryptoError>

Context-aware opening operations.

pub fn psk_seal( pubkey_r: &S::PublicKey, info: &[u8], psk: &[u8], psk_id: &[u8], plaintext: &[u8], aad: &[u8], out: &mut [u8], ) -> Result<S::EncapsulatedKey, UnknownCryptoError>

Available on crate feature safe_api only.

One-shot API for HPKE Psk mode Self::seal() operation.

pub fn psk_open( enc: &S::EncapsulatedKey, secret_key_r: &S::PrivateKey, info: &[u8], psk: &[u8], psk_id: &[u8], ciphertext: &[u8], aad: &[u8], out: &mut [u8], ) -> Result<(), UnknownCryptoError>

One-shot API for HPKE Psk mode Self::open() operation.

pub fn export_secret( &self, exporter_context: &[u8], out: &mut [u8], ) -> Result<(), UnknownCryptoError>

Export secret.